# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Akasha API",
    "description": "Personal knowledge retrieval.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Your collections, by name.",
        "operationId": "list_collections",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionList"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create a collection, optionally with files in it.",
        "operationId": "create_collection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      }
    },
    "/api/v1/collections/{id}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "One of your collections.",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete a collection. Its files are kept.",
        "operationId": "delete_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "collections"
        ],
        "summary": "Rename, describe, recolour or change the icon of a collection.",
        "operationId": "update_collection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      }
    },
    "/api/v1/collections/{id}/files": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Add files to a collection (up to 100 per request). Files already in it, and\nids that are not your files, are skipped.",
        "operationId": "add_collection_files",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionFilesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionFilesChanged"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      }
    },
    "/api/v1/collections/{id}/files/remove": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Take files out of a collection (the files themselves are kept).",
        "operationId": "remove_collection_files",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Collection id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionFilesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionFilesChanged"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionColor": {
        "type": "string",
        "description": "A collection's colour, from the UI's muted palette (never a raw colour).",
        "enum": [
          "sage",
          "sky",
          "ochre",
          "clay",
          "plum",
          "slate"
        ]
      },
      "CollectionFilesChanged": {
        "type": "object",
        "required": [
          "file_ids",
          "collection"
        ],
        "properties": {
          "collection": {
            "$ref": "#/components/schemas/CollectionResponse",
            "description": "The collection afterwards."
          },
          "file_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The files actually added (or removed); already-present or unknown ids are left out."
          }
        }
      },
      "CollectionFilesRequest": {
        "type": "object",
        "required": [
          "file_ids"
        ],
        "properties": {
          "file_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Up to 100 file ids. Ids that are not your files are skipped."
          }
        }
      },
      "CollectionIcon": {
        "type": "string",
        "description": "A collection's icon.",
        "enum": [
          "folder",
          "book",
          "briefcase",
          "flask",
          "heart",
          "star",
          "archive",
          "receipt",
          "plane",
          "home",
          "graduation",
          "code"
        ]
      },
      "CollectionList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionResponse"
            },
            "description": "By name."
          }
        }
      },
      "CollectionResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "color",
          "icon",
          "file_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/CollectionColor"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "Plain text, may be empty."
          },
          "file_count": {
            "type": "integer",
            "format": "int64"
          },
          "icon": {
            "$ref": "#/components/schemas/CollectionIcon"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last change, including files added or removed."
          }
        }
      },
...
      "CreateCollection": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CollectionColor"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Up to 2000 characters."
          },
          "file_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Files to put in it right away (up to 100; ids that are not yours are skipped)."
          },
          "icon": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CollectionIcon"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "1-100 characters, unique among your collections (ignoring case)."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateCollection": {
        "type": "object",
        "properties": {
          "color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CollectionColor"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"\"` clears it."
          },
          "icon": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CollectionIcon"
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal API token (akasha_pat_…)"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "akasha_session"
      }
    }
  }
}